move contract definitions to primitives
